requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-docx==1.1.0
pdfminer.six==20231228
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
PyMuPDF==1.24.10
python-multipart==0.0.9
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
scikit-learn==1.3.2
joblib==1.3.2
slowapi==0.1.9
pymysql==1.1.0
